pyproject.toml
[tool.poetry]
name = "simple-poop"
version = "0.1.0"
description = ""
authors = ["Max <realfalf@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
aiogram = "^3.15.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
